+++ elasticsearch-base/src/main/java/org/vertexium/elasticsearch/ElasticSearchSearchIndexBase.java
@@ -44,6 +44,7 @@ public abstract class ElasticSearchSearchIndexBase implements SearchIndex, Searc
     public static final String VISIBILITY_FIELD_NAME = "__visibility";
     public static final String OUT_VERTEX_ID_FIELD_NAME = "__outVertexId";
     public static final String IN_VERTEX_ID_FIELD_NAME = "__inVertexId";
+    public static final String EDGE_LABEL_FIELD_NAME = "__edgeLabel";
     public static final String EXACT_MATCH_PROPERTY_NAME_SUFFIX = "_e";
     public static final String GEO_PROPERTY_NAME_SUFFIX = "_g";
     public static final int MAX_BATCH_COUNT = 25000;
@@ -261,6 +262,7 @@ public abstract class ElasticSearchSearchIndexBase implements SearchIndex, Searc
                 .startObject(VISIBILITY_FIELD_NAME).field("type", "string").field("analyzer", "keyword").field("index", "not_analyzed").field("store", "true").endObject()
                 .startObject(IN_VERTEX_ID_FIELD_NAME).field("type", "string").field("analyzer", "keyword").field("index", "not_analyzed").field("store", "true").endObject()
                 .startObject(OUT_VERTEX_ID_FIELD_NAME).field("type", "string").field("analyzer", "keyword").field("index", "not_analyzed").field("store", "true").endObject()
+                .startObject(EDGE_LABEL_FIELD_NAME).field("type", "string").field("analyzer", "keyword").field("index", "not_analyzed").field("store", "true").endObject()
         ;
         getConfig().getScoringStrategy().addFieldsToElementType(builder);
     }